@@ -8,13 +8,13 @@ import { LoginComponent } from './login';
 
 
 
-import {MatFormFieldModule} from '@angular/material/form-field';
-import {MatIconModule} from '@angular/material/icon';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatIconModule } from '@angular/material/icon';
 
 import { BrowserModule } from '@angular/platform-browser';
 import { ReactiveFormsModule } from '@angular/forms';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
-import { MatInputModule} from '@angular/material/input';
+import { MatInputModule } from '@angular/material/input';
 
 // used to create fake backend
 import { fakeBackendProvider } from './_helpers';
@@ -24,8 +24,8 @@ import { fakeBackendProvider } from './_helpers';
 import { JwtInterceptor, ErrorInterceptor } from './_helpers';
 
 import { SocialLoginModule } from 'angularx-social-login';
-import { AuthServiceConfig, GoogleLoginProvider} from 'angularx-social-login';
-import {MatButtonModule} from '@angular/material/button';
+import { AuthServiceConfig, GoogleLoginProvider } from 'angularx-social-login';
+import { MatButtonModule } from '@angular/material/button';
 import { DemoMaterialModule } from '../material.module';
 import { ToastrModule } from 'ngx-toastr';
 
@@ -49,29 +49,31 @@ export function provideConfig() {
 
 
 @NgModule({
-  declarations: [HeaderComponent,HomeComponent,LoginComponent],
+  declarations: [HeaderComponent, HomeComponent, LoginComponent],
   imports: [
     CommonModule,
     CoreRoutingModule,
     BrowserModule,
-        ReactiveFormsModule,
-        HttpClientModule,
-        MatFormFieldModule,
-        MatIconModule,
-        SocialLoginModule,
-        MatInputModule,
-        MatButtonModule,
-        DemoMaterialModule,
-        ToastrModule.forRoot()
-        
+    ReactiveFormsModule,
+    HttpClientModule,
+    MatFormFieldModule,
+    MatIconModule,
+    SocialLoginModule,
+    MatInputModule,
+    MatButtonModule,
+    DemoMaterialModule,
+    ToastrModule.forRoot()
+
   ], providers: [
     { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true },
     { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
-    {provide: AuthServiceConfig,
-      useFactory: provideConfig},
+    {
+      provide: AuthServiceConfig,
+      useFactory: provideConfig
+    },
 
     // provider used to create fake backend
     fakeBackendProvider
-],exports:[HeaderComponent]
+  ], exports: [HeaderComponent]
 })
 export class CoreModule { }
