@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { AdminRoutingModule } from './admin-routing.module';
-import {MatSelectModule} from '@angular/material/select';
+import { MatSelectModule } from '@angular/material/select';
 
 import { AdminFlightsComponent } from './admin-flights/admin-flights.component';
 import { HttpClientModule } from '@angular/common/http';
@@ -13,7 +13,7 @@ import { EditCustomerComponent } from './edit-customer/edit-customer.component';
 import { DemoMaterialModule } from '../material.module';
 import { AdminFlightPipe } from './admin-flight.pipe';
 import { NgxSpinnerModule } from 'ngx-spinner';
-import { MatDialogRef} from '@angular/material/dialog';
+import { MatDialogRef } from '@angular/material/dialog';
 @NgModule({
   declarations: [AdminFlightsComponent, SelectedFlightComponent, EditCustomerComponent, AdminFlightPipe],
   imports: [
@@ -21,14 +21,15 @@ import { MatDialogRef} from '@angular/material/dialog';
     AdminRoutingModule,
     HttpClientModule,
     DemoMaterialModule,
-    NgxSpinnerModule
-   
-  ], exports:[AdminFlightPipe],
+    NgxSpinnerModule,
+
+
+  ], exports: [AdminFlightPipe],
   providers: [
     {
       provide: MatDialogRef,
       useValue: {}
     }
- ]
+  ]
 })
 export class AdminModule { }
