@@ -3,19 +3,10 @@ import { Component, OnInit, Optional, Inject } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 
 import { AdminService } from '../admin.service';
+import { Passenger } from 'src/app/core/_models/passenger';
+import { Flight } from 'src/app/core/_models/flight';
 
-export interface UsersData {
-  cId: number,
-  customername: string,
-  seatno: number,
-  checkenin: boolean,
-  passport: string,
-  address: string,
-  DOB: string,
-  category: string,
-  ancillayService: string[]
 
-}
 
 @Component({
   selector: 'app-edit-customer',
@@ -27,25 +18,27 @@ export class EditCustomerComponent implements OnInit {
   local_data: any;
   categories: string[] = ['wheel', 'infants', 'normal'];
   ancillaryServices: string[];
-  allPassenger:any;
-  totalSeat=["1A", "1B", "1C", "1D", "1E", "1F","2A", "2B", "2C", "2D", "2E", "2F",
-"3A", "3B", "3C", "3D", "3E", "3F",
-"4A", "4B", "4C", "4D", "4E", "4F",
-"5A", "5B", "5C", "5D", "5E", "5F",
-"6A", "6B", "6C", "6D", "6E", "6F",
-"7A", "7B", "7C", "7D", "7E", "7F",
-"8A", "8B", "8C", "8D", "8E", "8F",
-"9A", "9B", "9C", "9D", "9E", "9F"];
+  allPassenger: Passenger[];
 
-bookedSeat:String[]=[];
-specialMeals:string[];
-shoppingItems:string[];
-reaminingSeat=this.totalSeat;
+
+  totalSeat = ["1A", "1B", "1C", "1D", "1E", "1F", "2A", "2B", "2C", "2D", "2E", "2F",
+    "3A", "3B", "3C", "3D", "3E", "3F",
+    "4A", "4B", "4C", "4D", "4E", "4F",
+    "5A", "5B", "5C", "5D", "5E", "5F",
+    "6A", "6B", "6C", "6D", "6E", "6F",
+    "7A", "7B", "7C", "7D", "7E", "7F",
+    "8A", "8B", "8C", "8D", "8E", "8F",
+    "9A", "9B", "9C", "9D", "9E", "9F"];
+
+  bookedSeat: String[] = [];
+  specialMeals: string[];
+  shoppingItems: string[];
+  reaminingSeat = this.totalSeat;
 
 
 
   constructor(public dialogRef: MatDialogRef<EditCustomerComponent>,
-    @Optional() @Inject(MAT_DIALOG_DATA) public data: UsersData,
+    @Optional() @Inject(MAT_DIALOG_DATA) public data: Passenger,
     private service: AdminService) {
     console.log(data);
     this.local_data = { ...data };
@@ -53,25 +46,24 @@ reaminingSeat=this.totalSeat;
   }
 
 
-  flight: any;
+  flight: Flight;
   ngOnInit(): void {
     this.flight = this.service.getSelectedFlight();
-    
-    
-    this.ancillaryServices = this.flight?.ancillaryServices;
-   console.log(this.ancillaryServices);
-   this.specialMeals=this.flight?.specialMeals;
-    this.shoppingItems=this.flight?.shoppingItems;
-    this.allPassenger=this.flight?.passengers;
-    let y=this.countBookedSeat()
+    this.ancillaryServices = this.flight.ancillaryServices;
+    this.specialMeals = this.flight.specialMeals;
+    this.shoppingItems = this.flight.shoppingItems;
+    this.allPassenger = this.flight.passengers;
+    let y = this.countBookedSeat()
 
   }
-  
-  doAction() {
-    console.log(this.local_data);
 
+
+
+  doAction() {
     this.dialogRef.close({ event: this.action, data: this.local_data });
   }
+
+
 
   closeDialog() {
     this.dialogRef.close({ event: 'Cancel' });
@@ -79,25 +71,16 @@ reaminingSeat=this.totalSeat;
 
 
   countBookedSeat() {
-   
-    
-    this.allPassenger?.find((a)=>{
-      this.totalSeat?.find((b)=>{
-        if(a.seatno===b)
-        {
+    this.allPassenger.find((a) => {
+      this.totalSeat.find((b) => {
+        if (a.seatno === b) {
           const index: number = this.reaminingSeat.indexOf(b);
-         
-          
-    if (index !== -1) {
-        this.reaminingSeat.splice(index, 1);
-    }        
-       
-          
+          if (index !== -1) {
+            this.reaminingSeat.splice(index, 1);
+          }
         }
 
       })
     })
-    
   }
-
 }
