@@ -7,17 +7,17 @@ import { CheckinService } from '../../checkin.service';
 import { AnyARecord } from 'dns';
 import { Passenger } from 'src/app/core/_models/passenger';
 import { Flight } from 'src/app/core/_models/flight';
+import { of } from 'rxjs';
 
 describe('PassengerDetailComponent', () => {
   let component: PassengerDetailComponent;
   let fixture: ComponentFixture<PassengerDetailComponent>;
-  let testService:CheckinService;
-  let checkpassenger:Passenger[];
-  let allPassenger:Passenger[];
-  let passengers:Passenger[];
-  
-  let flight:Flight;
-  let totalSeat:string[];
+  let dataService: CheckinService;
+
+  let passengers: Passenger[];
+
+  let flight: Flight;
+  let spy2: any;
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       imports: [RouterTestingModule, HttpClientTestingModule],
@@ -30,134 +30,160 @@ describe('PassengerDetailComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(PassengerDetailComponent);
     component = fixture.componentInstance;
-   flight={
-    "id": 1,
-    "planename": "Indigo",
-    "planeno": "6E-2131",
-    "departure": "Mon May 25 2020 11:48:22 GMT+0530 (India Standard Time)",
-    "departureFrom": "DELHI",
-    "arrivalTo": "HYDERABAD",
-    "arrival": "Mon May 25 2020 08:48:22 GMT+0530 (India Standard Time)",
-    "duration": "02 hrs",
-    "price": 2341,
-    "ancillaryServices": [
-      "ancillary1",
-      "ancillary2 ",
-      "ancillary3 ",
-      "anci4"
-    ],
-    "specialMeals": [
-      "Normal Meal",
-      "Special Meal"
-    ],
-    "shoppingItems": [
-      "choclate",
-      "lays"
-    ],
-    "image": "https://images-eu.ssl-images-amazon.com/images/I/418wwCw2HuL.png",
-    "passengers": [
-      {
-        "cId": 13,
-        "customername": "samir",
-        "checkenin": true,
-        "passport": "",
-        "address": "Bellagam123",
-        "DOB": "1998-06-19T18:30:00.000Z",
-        "category": "infants",
-        "ancillayService": [
-          "ancillary1",
-          "ancillary2 ",
-          "ancillary3 "
-        ],
-        "action": "Update",
-        "seatno": "2A",
-        "meal": "Special Meal",
-        "shop": [
-          "choclate",
-          "lays"
-        ]
-      }]}
-      passengers=flight.passengers;
-      allPassenger=flight.passengers;
-      totalSeat=["5D","6F"]
+    dataService = fixture.debugElement.injector.get(CheckinService);
+    flight = {
+      "id": 1,
+      "planename": "Indigo",
+      "planeno": "6E-2131",
+      "departure": "Mon May 25 2020 11:48:22 GMT+0530 (India Standard Time)",
+      "departureFrom": "DELHI",
+      "arrivalTo": "HYDERABAD",
+      "arrival": "Mon May 25 2020 08:48:22 GMT+0530 (India Standard Time)",
+      "duration": "02 hrs",
+      "price": 2341,
+      "ancillaryServices": [
+        "ancillary1",
+        "ancillary2 ",
+        "ancillary3 ",
+        "anci4"
+      ],
+      "specialMeals": [
+        "Normal Meal",
+        "Special Meal"
+      ],
+      "shoppingItems": [
+        "choclate",
+        "lays"
+      ],
+      "image": "https://images-eu.ssl-images-amazon.com/images/I/418wwCw2HuL.png",
+      "passengers": [
+        {
+          "cId": 13,
+          "customername": "samir",
+          "checkenin": true,
+          "passport": "",
+          "address": "Bellagam123",
+          "DOB": "1998-06-19T18:30:00.000Z",
+          "category": "infants",
+          "ancillayService": [
+            "ancillary1",
+            "ancillary2 ",
+            "ancillary3 "
+          ],
+          "action": "Update",
+          "seatno": "2A",
+          "meal": "Special Meal",
+          "shop": [
+            "choclate",
+            "lays"
+          ]
+        }]
+    }
+    passengers = flight.passengers;
+    let spy = spyOn(dataService, 'getSelectedPassenger')
+      .and.returnValue(passengers[0]);
+    let spy1 = spyOn(dataService, 'getSelectedFlight')
+      .and.returnValue(flight);
+    spy2 = spyOn(dataService, 'update')
+      .and.returnValue(of(flight));
     fixture.detectChanges();
   });
 
+
   it('should create', () => {
-    console.log(flight);
-    
     expect(component).toBeTruthy();
   });
 
 
-  // it('should call the GetAddressList of checkout service', function() {
-   
-  //   spyOn(testService, 'getSelectedPassenger').and.callThrough();
-  //   // testAdminService.GetRequestedList();
-  //   component.ngOnInit();
-  //   expect(testService.getSelectedPassenger).toHaveBeenCalled();
 
- 
-
-  // });
+  it('should call service', async(() => {
+    fixture.whenStable().then(() => {
+      expect(component.flight.planeno).toBe(flight.planeno)
+    })
+  }))
 
 
-  it('should call AddAddress of chechout add', function() { 
-     component.updateTable();
-   });
+
+  it('should call service', async(() => {
+    fixture.whenStable().then(() => {
+      expect(component.passenger.customername).toBe(passengers[0].customername)
+    })
+  }))
 
 
-  it('should call AddAddress of chechout add', function () {
+  it('should call updateTable', async(() => {
+    fixture.whenStable().then(() => {
+      expect(component.passenger.customername).toBe(passengers[0].customername)
+    })
+    component.updateTable();
+    expect(spy2).toHaveBeenCalled();
+  }));
 
-    let array = ["sam", "ram"];
-    let id: Number;
-    component.findIndex(array, id);
+
+  it('should call findIndex', function () {
+    component.flight = flight;
+    component.findIndex(passengers, "samir", 13);
+    expect(component.findIndex(passengers, "samir", 13)).toBe(0);
   });
 
-  //  it('should call AddAddress of chechout add', function() {
-  //    component.checkIn();
-  //  });
+
+  it('should call countBookedSeat', function () {
+    component.countBookedSeat();
+  });
 
 
-  //  it('should call AddAddress of chechout add', function() {
-  //    component.changeSeat();
-  //  });
+  it('should call checkSeat', function () {
+    component.isChangeSeat = true;
+    component.checkSeat();
+    expect(component.isChangeSeat).toBe(false);
+  });
 
 
-   it('should call AddAddress of chechout add', function() {
-    
-     component.countBookedSeat();
-   });
-
-
-   it('should call AddAddress of chechout add', function() {
-     component.checkSeat();
-   });
-   it('should call AddAddress of chechout add', function() {
-     component.changeCancel();
-   });
+  it('should call changeCancel', function () {
+    component.isChangeSeat = true;
+    component.changeCancel();
+    expect(component.isChangeSeat).toBe(false);
+  });
 
 
 
-   it('should call AddAddress of chechout add', function() {
-      let data2:any;
-     component.getSelected(data2);
-   });
+  it('should call getSelected', function () {
+    let data2: any;
+    component.getSelected(data2);
+  });
 
 
-   it('should call AddAddress of chechout add', function() {
-    let flight={
-      passengers:[]
-          };
-      let data2=flight.passengers;
-     component.getTrial(data2);
-   });
+  it('should call getTrial true', function () {
+    component.passenger = passengers[0];
+    let data2 = ["2A"];
+    component.getTrial(data2);
+    expect(component.isCheckIn).toBe(true);
+  });
 
-   it('should call check of chechout add', function() {    
-   
-     checkpassenger=flight.passengers;
-     component.check(checkpassenger);
-   });
+
+  it('should call getTrial flase', function () {
+    component.passenger = passengers[0];
+    let data2 = [];
+    component.getTrial(data2);
+    expect(component.isCheckIn).toBe(false);
+  });
+
+
+  it('should call check of chechout add', function () {
+    component.check(passengers);
+  });
+
+
+  it('should call check of chechout add', function () {
+    component.checkIn();
+  });
+
+
+  it('should call changeSeat', function () {
+    component.isChangeSeat = true;
+    component.changeSeat();
+    expect(component.isChangeSeat).toBe(false);
+  });
 
 
 
