@@ -26,7 +26,6 @@ export class AuthenticationService {
     login(username: string, password: string) {
         return this.http.post<any>(`${environment.apiUrl}/users/authenticate`, { username, password })
             .pipe(map(user => {
-                console.log(user);
                 
                 // login successful if there's a jwt token in the response
                 if (user && user.token) {
@@ -51,7 +50,7 @@ export class AuthenticationService {
         
         localStorage.setItem('currentUser', JSON.stringify(this.socialLoggedUsed));
         this.currentUserSubject.next(this.socialLoggedUsed);
-        console.log(this.currentUserSubject.value);
+       
         
         return socialUser;
     }
